Refactor session into module
